completed tiles and lists
